pyproject.toml
[tool.poetry]
name = "flask-blog"
version = "0.1.0"
description = "MLH Production Engineering Track Project"
authors = ["cjlaserna <49669196+cjlaserna@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
cffi = "1.15.0"
click = "8.0.1"
cryptography = "37.0.2"
Flask = "2.0.1"
itsdangerous = "2.0.1"
Jinja2 = "3.0.1"
MarkupSafe = "2.0.1"
peewee = "3.14.10"
pycparser = "2.21"
PyMySQL = "1.0.2"
python-dotenv = "0.17.1"
Werkzeug = "2.0.1"
gql = {extras = ["all"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
